Improve appearance of modal
